pyproject.toml
[tool.poetry]
name = "data-challenge"
version = "0.1.0"
description = ""
authors = ["Daniel Montoya <carlos.danmontoya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
jupyter = "^1.1.1"
pandas-gbq = "^0.24.0"
google-cloud-bigquery = "^3.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
